requirements.txt
django
django-redis
djangorestframework
requests